+++ gridcapa-core-cc-api/src/main/java/com/farao_community/farao/gridcapa_core_cc/api/exception/CoreCCRaoException.java
@@ -1,9 +1,8 @@
 /*
- * Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
  *  This Source Code Form is subject to the terms of the Mozilla Public
  *  License, v. 2.0. If a copy of the MPL was not distributed with this
  *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
- *
  */
 
 package com.farao_community.farao.gridcapa_core_cc.api.exception;
